Homepage: Fixed field_group config export for html_element type group.
The configuration for the group_email field group of the
localgov_subscribe_panel Paragraph had two unintended properties in it: "open",
and "description".  It appears that these have been added due to a field_group
module bug where a group is created as a "Detail" item and then changed to an
"Html element".

@see https://github.com/localgovdrupal/localgov/issues/107